requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
numpy==1.26.4
pillow==11.1.0
psutil==7.0.0
scikit-image==0.25.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
